Fix inconsistency with libc attribute decoration
Decorating the `Event` structure with the "packed" attribute,
regardless of the platform, is incorrect since libc decorates
the`libc::epoll_event` only for x86_64 platforms.
As a consequence, on other architectures the size of the
`Event` structure is different than that of the
`libc::epoll_event` which is a trigger for inconsistent behavior.

Signed-off-by: Diana Popa <dpopa@amazon.com>